added getCalendarEvents function
The code decodes the provided access token, sets it in the OAuth2 client, and makes a request to the Google Calendar API to retrieve the events. The response is then formatted and returned as a JSON object. If there's an error during the process, an error response is returned instead.